Move DefaultChecker and DeviceConformanceTests to src/python_testing
These shared modules are used by production TC files in src/python_testing/
and belong at that level rather than in the test_testing/ subdirectory.
Update all import sites accordingly and add sys.path fixups in the
test_testing/ unit tests that now need to reach the parent directory.
