+++ src/python_testing/TC_DeviceBasicComposition.py
@@ -193,7 +193,7 @@ import os
 from dataclasses import dataclass
 from typing import Any, Callable
 
-from test_testing.DeviceConformanceTests import get_supersets
+from DeviceConformanceTests import get_supersets
 
 import matter.clusters as Clusters
 import matter.clusters.ClusterObjects
