+++ src/python_testing/test_testing/TestDefaultWarnings.py
@@ -15,8 +15,11 @@
 #    limitations under the License.
 #
 
+import os
+import sys
 from typing import Callable
 
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 import DefaultChecker as DefaultChecker
 from mobly import asserts
