+++ src/python_testing/test_testing/TestSpecParsingSelection.py
@@ -15,7 +15,10 @@
 #    limitations under the License.
 #
 
+import os
+import sys
 
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 from DeviceConformanceTests import DeviceConformanceTests
 from mobly import asserts, signals
